# apps/coreApi/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/backlog/tasks": {
      "get": {
        "operationId": "TaskController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tasks with pagination and filters",
        "tags": [
          "backlog"
        ]
      },
      "post": {
        "operationId": "TaskController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new task",
        "tags": [
          "backlog"
        ]
      }
    },
    "/api/v1/backlog/tasks/{id}": {
      "get": {
        "operationId": "TaskController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get task by ID",
        "tags": [
          "backlog"
        ]
      },
      "patch": {
        "operationId": "TaskController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a task",
        "tags": [
          "backlog"
        ]
      },
      "delete": {
        "operationId": "TaskController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a task",
        "tags": [
          "backlog"
        ]
      }
    },
    "/api/v1/backlog/sprints": {
      "get": {
        "operationId": "SprintController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List sprints with pagination and filters",
        "tags": [
          "backlog"
        ]
      },
      "post": {
        "operationId": "SprintController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new sprint",
        "tags": [
          "backlog"
        ]
      }
    },
    "/api/v1/backlog/sprints/{id}": {
      "get": {
        "operationId": "SprintController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get sprint by ID",
        "tags": [
          "backlog"
        ]
      },
      "patch": {
        "operationId": "SprintController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a sprint",
        "tags": [
          "backlog"
        ]
      },
      "delete": {
        "operationId": "SprintController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a sprint",
        "tags": [
          "backlog"
        ]
      }
    },
    "/api/v1/backlog/sprints/{id}/tasks": {
      "post": {
        "operationId": "SprintController_assignTasks",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign tasks to a sprint",
        "tags": [
          "backlog"
        ]
      }
    },
...
  },
  "info": {
    "title": "nodeAdmin API",
    "description": "Enterprise multi-tenant SaaS middleware platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}